add environment banner
