+++ src/components/navigation/Panel.tsx
@@ -1,11 +1,10 @@
-import { Trans, useTransContext } from "@mbarzda/solid-i18next";
-import { A, useLocation } from "@solidjs/router";
-import { JSX, Show, createEffect, createSignal } from "solid-js";
+import { useLocation } from "@solidjs/router";
+import { JSX, createEffect } from "solid-js";
 
-import { BurgerArrowIcon, BurgerIcon, CloseIcon } from "../../components/icons";
+import { BurgerArrowIcon, BurgerIcon } from "../../components/icons";
 import { getSlots } from "../../components/layout/Slot";
 import { clickOutside } from "../../directives";
-import { TKEYS } from "../../locales";
+import { EnvironmentBanner } from "../content/EnvironmentBanner";
 import styles from "./Panel.module.scss";
 
 false && clickOutside;
@@ -18,15 +17,10 @@ type Props = {
 };
 
 export function Panel(props: Props) {
-  const [trans] = useTransContext();
-
   const location = useLocation();
 
   /* eslint-disable-next-line */
   const slots = getSlots(props.children);
-
-  // const [showSlider, setShowSlider] = createSignal(false);
-  const [showEnvironmentBanner, setShowEnvironmentBanner] = createSignal(true);
 
   createEffect(() => {
     if (props.showSlider()) {
@@ -48,65 +42,44 @@ export function Panel(props: Props) {
     props.setShowSlider(false);
   }
 
-  function handleCloseBanner() {
-    setShowEnvironmentBanner(false);
-  }
-
   return (
-    <div style={props.style}>
-      <div class={styles.Panel}>
-        <BurgerIcon class={styles.MenuIcon} onClick={handleToggleSlider} />
+    <>
+      <div style={props.style}>
+        <div class={styles.Panel}>
+          <BurgerIcon class={styles.MenuIcon} onClick={handleToggleSlider} />
 
-        <div class={styles.Main}>{slots.logo}</div>
+          <div class={styles.Main}>{slots.logo}</div>
+        </div>
+
+        <div
+          class={styles.SliderBackground}
+          classList={{
+            [styles.SliderBackgroundIn]: props.showSlider(),
+            [styles.SliderBackgroundOut]: !props.showSlider(),
+          }}
+        />
+        <nav
+          class={styles.Slider}
+          classList={{
+            [styles.SlideIn]: props.showSlider(),
+            [styles.SlideOut]: !props.showSlider(),
+          }}
+          use:clickOutside={() => handleCloseSlider()}
+        >
+          <div class={styles.Menu}>
+            <BurgerArrowIcon
+              class={styles.MenuIcon}
+              onClick={handleCloseSlider}
+            />
+          </div>
+
+          <div class={styles.MainNavigation}>{slots.items}</div>
+
+          <div class={styles.Settings}>{slots.settings}</div>
+        </nav>
       </div>
 
-      <div
-        class={styles.SliderBackground}
-        classList={{
-          [styles.SliderBackgroundIn]: props.showSlider(),
-          [styles.SliderBackgroundOut]: !props.showSlider(),
-        }}
-      />
-      <nav
-        class={styles.Slider}
-        classList={{
-          [styles.SlideIn]: props.showSlider(),
-          [styles.SlideOut]: !props.showSlider(),
-        }}
-        use:clickOutside={() => handleCloseSlider()}
-      >
-        <div class={styles.Menu}>
-          <BurgerArrowIcon
-            class={styles.MenuIcon}
-            onClick={handleCloseSlider}
-          />
-        </div>
-
-        <div class={styles.MainNavigation}>{slots.items}</div>
-
-        <div class={styles.Settings}>{slots.settings}</div>
-      </nav>
-
-      <Show
-        when={
-          !import.meta.env.VITE_ENVIRONMENT?.startsWith("prod") &&
-          showEnvironmentBanner()
-        }
-      >
-        <div class={styles.EnvironmentBanner}>
-          <CloseIcon onClick={handleCloseBanner} />
-
-          <p>
-            <Trans key={TKEYS["environment-banner"].title} />
-          </p>
-          <span>
-            <Trans key={TKEYS["environment-banner"].description} />
-            <A href={trans(TKEYS.peoplesmarkets_main_link)}>
-              <Trans key={TKEYS.peoplesmarkets_main_link} />
-            </A>
-          </span>
-        </div>
-      </Show>
-    </div>
+      <EnvironmentBanner />
+    </>
   );
 }
